requirement_analysis_agent/requirements.txt
spacy
python-dotenv
pika
